backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
scikit-learn
numpy
requests
psycopg2-binary
sentence-transformers
jsonschema
httpx
pytest


